Refactor service worker event listeners into own file
